refactor: add validation when deleting service

+++ controllers/admin.controller.ts
@@ -360,8 +360,12 @@ class AdminController {
   static async deleteService(req: Request, res: Response) {
     try {
       const id = parseInt(req.params.id as string, 10);
-      await ServiceModel.deleteService(id);
+      const portfolios = await PortfolioModel.getAllPortfolios();
+      const isServiceUsed = portfolios.some((portfolio) => portfolio.serviceId === id);
 
+      if (isServiceUsed) throw new Error('Service is being used by portfolios, delete them first!');
+
+      await ServiceModel.deleteService(id);
       req.flash('alertMessage', 'Service deleted successfully');
       req.flash('alertType', 'success');
     } catch (error: any) {
